test: split research funding test to satisfy domain test file size gate (Refs #3512)
The repo.domain_package_test_file_size lint rule (<=400 physical lines)
failed because research_phase_funding_test.dart grew to 513 lines after
adding the Medium-pacing rebalance group. Move that group into a new
research_phase_funding_pacing_test.dart so both files stay under the cap.
